requirements.txt
# Core dependencies
azure-identity>=1.15.0
azure-storage-blob>=12.19.0
azure-search-documents>=11.4.0
openai>=1.12.0
msgraph-sdk>=1.2.0

# LangChain
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-text-splitters>=0.3.0

# Vector database
faiss-cpu>=1.7.4

# Document processing
python-pptx>=0.6.23
tiktoken>=0.5.0

# Local embeddings
scikit-learn>=1.4.0

# Web search
duckduckgo_search>=8.0.0

# Web framework
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0

# Redis for real-time transcript buffer
redis>=5.0.0

# Transcription services (optional)
assemblyai>=0.30.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.26.0
numpy>=1.26.0
tqdm>=4.66.0

# Security & Privacy
cryptography>=42.0.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.1.0
ruff>=0.2.0
mypy>=1.8.0
